+++ baseapp/baseapp_test.go
@@ -150,10 +150,12 @@ func TestInitChainer(t *testing.T) {
 	name := t.Name()
 	app := NewBaseApp(name, logger, db)
 
-	// make a cap key and mount the store
+	// make cap keys and mount the stores
+	// NOTE/TODO: mounting multiple stores is broken
+	// see https://github.com/cosmos/cosmos-sdk/issues/532
 	capKey := sdk.NewKVStoreKey("main")
-	capKey2 := sdk.NewKVStoreKey("key2")
-	app.MountStoresIAVL(capKey, capKey2)
+	// capKey2 := sdk.NewKVStoreKey("key2")
+	app.MountStoresIAVL(capKey)          // , capKey2)
 	err := app.LoadLatestVersion(capKey) // needed to make stores non-nil
 	assert.Nil(t, err)
 
@@ -186,8 +188,8 @@ func TestInitChainer(t *testing.T) {
 	// reload app
 	app = NewBaseApp(name, logger, db)
 	capKey = sdk.NewKVStoreKey("main")
-	capKey2 = sdk.NewKVStoreKey("key2")
-	app.MountStoresIAVL(capKey, capKey2)
+	// capKey2 = sdk.NewKVStoreKey("key2") // TODO
+	app.MountStoresIAVL(capKey)         //, capKey2)
 	err = app.LoadLatestVersion(capKey) // needed to make stores non-nil
 	assert.Nil(t, err)
 	app.SetInitChainer(initChainer)